requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==1.10.13
python-dotenv==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
email-validator==2.1.0.post1
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
numpy==1.26.2
scikit-learn 